requirements.txt
Flask<0.11,>=0.10.1
PyInstaller==3.1
coloredlogs==5.0
jsonschema==2.5
mitmproxy==0.17
pyYAML==3.11
pytest==2.9.2
pytoml==0.1
tinydb==3.1
